# artifacts/pos-system/src/backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Restaurant ERP + POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List all employees
      parameters:
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: role
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ===== REPORTS =====
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      parameters:
        - name: from
          in: query
          schema:
            type: ["string", "null"]
        - name: to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
...
components:
  schemas:
...
    Employee:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, cashier, waiter, chef, driver]
        salary:
          type: ["number", "null"]
        isActive:
          type: boolean
        joinDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - isActive
        - joinDate
        - createdAt
    EmployeeInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, cashier, waiter, chef, driver]
        salary:
          type: ["number", "null"]
        joinDate:
          type: string
      required:
        - name
        - role
        - joinDate
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, manager, cashier, waiter, chef, driver]
        salary:
          type: ["number", "null"]
        isActive:
          type: boolean
        joinDate:
          type: string

    # ===== REPORTS SCHEMAS =====
    SalesReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalDiscount:
          type: number
        totalTax:
          type: number
        averageOrderValue:
          type: number
        byPaymentMethod:
          type: array
          items:
            type: object
            properties:
              method:
                type: string
              count:
                type: integer
              total:
                type: number
            required:
              - method
              - count
              - total
        byOrderType:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              count:
                type: integer
              total:
                type: number
            required:
              - type
              - count
              - total
      required:
        - totalRevenue
        - totalOrders
        - totalDiscount
        - totalTax
        - averageOrderValue
        - byPaymentMethod
        - byOrderType
...